- Se termina el preprocesado. Unicamente faltaría añadir una barra de progreso. - Se modifica la clase MultiProcess para que los procesos se almacenen como smart pointers

+++ src/fme/widgets/RswheWidget.cpp
@@ -32,12 +32,12 @@ RswheWidget::~RswheWidget()
 
 void RswheWidget::onHistogramCutByMean()
 {
-  emit histogramCutChange(HistogramCut::by_mean);
+  emit histogramCutChange(static_cast<int>(HistogramCut::by_mean));
 }
 
 void RswheWidget::onHistogramCutByMedian()
 {
-  emit histogramCutChange(HistogramCut::by_median);
+  emit histogramCutChange(static_cast<int>(HistogramCut::by_median));
 }
 
 int RswheWidget::histogramDivisions() const
